bug sur lst_map memcmp et itoa ds maintest

+++ ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int		ft_strlen(const char *str)
+size_t		ft_strlen(const char *str)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (*str++)
